# fix [tools] Enhance build error handling for a build module named native