--- a/tools/src/main/scala/scala/scalanative/build/LLVM.scala
+++ b/tools/src/main/scala/scala/scalanative/build/LLVM.scala
@@ -173,12 +173,39 @@ private[scalanative] object LLVM {
       case Success(_) =>
     }
 
+  /** This function allows a project to have multiple `main` files by copying
+   *  the one selected to the same parent directory as the `workDir` which is by
+   *  default named `native`. Since the directory is named `native`, having a
+   *  project named `native` will by default produce a executable named `native`
+   *  which will throw an exception since the copy command uses
+   *  REPLACE_EXISTING.
+   *
+   *  Having a project or `baseName` named `native` conflict with the build.
+   */
   private def copyOutput(config: Config, buildPath: Path) = {
     val outPath = config.artifactPath
-    config.compilerConfig.buildTarget match {
-      case BuildTarget.Application =>
-        Files.copy(buildPath, outPath, StandardCopyOption.REPLACE_EXISTING)
-      case _: BuildTarget.Library => outPath
+    try {
+      config.compilerConfig.buildTarget match {
+        case BuildTarget.Application =>
+          Files.copy(buildPath, outPath, StandardCopyOption.REPLACE_EXISTING)
+        case _: BuildTarget.Library => outPath
+      }
+    } catch {
+      case t: Throwable =>
+        val msg =
+          if (outPath.toFile().exists())
+            s"""Executable build module or `baseName` is named 'native'
+                |which conflicts with the compiler `workDir`.
+                |Please rename the build module or
+                |use `withBaseName` to rename the executable.
+                |Cause: ${t}""".stripMargin
+          else
+            s"""Unexpected exception:
+                |Build path: $buildPath
+                |Artifact: $outPath
+                |Cause: ${t}""".stripMargin
+
+        throw new BuildException(msg)
     }
   }